🏗️ refactor: rename CliHub entry point module and update view header text
