+++ src/clihub/index.html
@@ -7,6 +7,6 @@
 	<link href="${styleUri}" rel="stylesheet">
 </head>
 <body>
-	<h1>hello CLI 3</h1>
+	<h1>hello CLI Hub</h1>
 </body>
 </html>
